+++ src/Thumbnail/index.types.ts
@@ -1,9 +1,9 @@
 // Types
 import type { GridProps } from '@mui/material'
 
-export type SpinoramaThumbnailDotProps = GridProps & {
+export type SpinoramaThumbnailProps = GridProps & {
 	/**
-	 * Indicates whether the dot is currently selected
+	 * Indicates whether the thumbnail is currently selected
 	 * Use this prop to style the component when it represents the selected thumbnail
 	 * @default false
 	 */